+++ frontend/src/app/core/model/training.model.ts
@@ -134,29 +134,3 @@ export interface TrainingRecordCategories {
   training: Training[];
   isMandatory: boolean;
 }
-
-export interface UploadCertificateSignedUrlRequest {
-  files: { filename: string }[];
-}
-
-export interface UploadCertificateSignedUrlResponse {
-  files: { filename: string; fileId: string; signedUrl: string; key: string }[];
-}
-
-export interface DownloadCertificateSignedUrlResponse {
-  files: { filename: string; signedUrl: string }[];
-}
-
-export interface S3UploadResponse {
-  headers: { etag: string };
-}
-export interface FileInfoWithETag {
-  filename: string;
-  fileId: string;
-  etag: string;
-  key: string;
-}
-
-export interface ConfirmUploadRequest {
-  files: { filename: string; fileId: string; etag: string }[];
-}
