+++ frontend/src/app/core/model/trainingAndQualifications.model.ts
@@ -33,3 +33,29 @@ export interface CertificateDownload {
 export type CertificateDownloadEvent = TrainingCertificateDownloadEvent | QualificationCertificateDownloadEvent;
 
 export type CertificateUploadEvent = TrainingCertificateUploadEvent | QualificationCertificateUploadEvent;
+
+export interface UploadCertificateSignedUrlRequest {
+  files: { filename: string }[];
+}
+
+export interface UploadCertificateSignedUrlResponse {
+  files: { filename: string; fileId: string; signedUrl: string; key: string }[];
+}
+
+export interface DownloadCertificateSignedUrlResponse {
+  files: { filename: string; signedUrl: string }[];
+}
+
+export interface S3UploadResponse {
+  headers: { etag: string };
+}
+export interface FileInfoWithETag {
+  filename: string;
+  fileId: string;
+  etag: string;
+  key: string;
+}
+
+export interface ConfirmUploadRequest {
+  files: { filename: string; fileId: string; etag: string }[];
+}
